Add nutrition goal update functionality and related tests
- Implemented the NutritionGoalUpdatedIntegrationEventConsumer to handle updates to nutrition goals based on integration events.
- Created unit tests for the consumer to ensure correct behavior when no overviews are found, when goals are updated, and when exceptions occur.
- Enhanced the GetDailyNutritionOverviewsByDateRange query to support new OverviewType and AggregationMode parameters.
- Updated related query handlers and validators to accommodate the new features.
- Added DTOs for daily nutrition overviews to facilitate data transfer.
- Refactored existing code to improve clarity and maintainability, including the introduction of new enums for overview types and aggregation modes.
- Updated functional tests to validate the new functionality and ensure comprehensive coverage.
- Enhanced error handling and logging for better traceability during event processing.

+++ Modules/Users/TrackYourLife.Modules.Users.Application/Core/Abstraction/Services/INutritionGoalsCalculator.cs
@@ -7,13 +7,13 @@ public interface INutritionGoalsCalculator
 {
     public int CalculateCalories(
         int age,
-        int weight,
-        int height,
+        float weight,
+        float height,
         Gender gender,
         ActivityLevel activityLevel,
         FitnessGoal FitnessGoal
     );
-    int CalculateProtein(int weight);
+    int CalculateProtein(float weight);
     int CalculateCarbs(int calories);
     int CalculateFat(int calories);
 }
